Frontend: Extract UI Events
Continue grouping GUI responsibilities into new package
forge.frontend...

forge.gui.events --> forge.frontend.events
