+++ forge-gui/src/main/java/forge/frontend/events/IUiEventVisitor.java
@@ -1,4 +1,4 @@
-package forge.gui.events;
+package forge.frontend.events;
 
 public interface IUiEventVisitor<T> {
     T visit(UiEventBlockerAssigned event);
