+++ forge-gui/src/main/java/forge/frontend/events/UiEventBlockerAssigned.java
@@ -1,4 +1,4 @@
-package forge.gui.events;
+package forge.frontend.events;
 
 import forge.game.card.CardView;
