+++ forge-gui/src/main/java/forge/gamemodes/match/HostedMatch.java
@@ -17,6 +17,11 @@ import com.google.common.collect.Maps;
 import com.google.common.eventbus.Subscribe;
 
 import forge.LobbyPlayer;
+import forge.frontend.events.IUiEventVisitor;
+import forge.frontend.events.UiEvent;
+import forge.frontend.events.UiEventAttackerDeclared;
+import forge.frontend.events.UiEventBlockerAssigned;
+import forge.frontend.events.UiEventNextGameDecision;
 import forge.game.Game;
 import forge.game.GameRules;
 import forge.game.GameType;
@@ -35,11 +40,6 @@ import forge.gui.GuiBase;
 import forge.gui.control.FControlGameEventHandler;
 import forge.gui.control.FControlGamePlayback;
 import forge.gui.control.WatchLocalGame;
-import forge.gui.events.IUiEventVisitor;
-import forge.gui.events.UiEvent;
-import forge.gui.events.UiEventAttackerDeclared;
-import forge.gui.events.UiEventBlockerAssigned;
-import forge.gui.events.UiEventNextGameDecision;
 import forge.gui.interfaces.IGuiGame;
 import forge.interfaces.IGameController;
 import forge.localinstance.properties.ForgeConstants;
