+++ forge-gui/src/main/java/forge/player/PlayerControllerHuman.java
@@ -53,6 +53,7 @@ import forge.deck.CardPool;
 import forge.deck.Deck;
 import forge.deck.DeckRecognizer;
 import forge.deck.DeckSection;
+import forge.frontend.events.UiEventNextGameDecision;
 import forge.game.Game;
 import forge.game.GameEntity;
 import forge.game.GameEntityView;
@@ -123,7 +124,6 @@ import forge.gamemodes.match.input.InputSelectEntitiesFromList;
 import forge.gui.FThreads;
 import forge.gui.GuiBase;
 import forge.gui.control.FControlGamePlayback;
-import forge.gui.events.UiEventNextGameDecision;
 import forge.gui.interfaces.IGuiGame;
 import forge.gui.util.SOptionPane;
 import forge.interfaces.IDevModeCheats;
